Massive code clean up - Also updated deps - Added new stuff for coding roles - Now it doesn't delete the invite if the sender was an admin - etc that I don't remember

+++ src/main/java/commands/myServer/AntiLink.java
@@ -1,30 +1,36 @@
 package commands.myServer;
 
+import main.Bumblebot;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageUpdateEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
 import utility.ConfigUtil;
+import utility.OtherUtil;
 import utility.core.UsrMsgUtil;
 
 public class AntiLink extends ListenerAdapter{
 
-	private String inviteRegex = "di?sco?rd(?:(\\.(?:me|io|gg|li)|sites\\.com)\\/.{0,4}|app\\.com.{1,4}(?:invite|oauth2).{0,5}\\/)\\w+";
+	private final String inviteRegex = "di?sco?rd(?:(\\.(?:me|io|gg|li)|sites\\.com).{0,4}|app\\.com.{1,4}(?:invite|oauth2).{0,5})\\w+";
 
 	
 	public void onGuildMessageReceived(GuildMessageReceivedEvent e) {
 		if(e.getGuild().getId().equals(ConfigUtil.getServerId())) {
-			if(e.getMessage().getContentRaw().replaceAll(this.inviteRegex, "").length() < e.getMessage().getContentRaw().length()) {
-				e.getChannel().deleteMessageById(e.getMessage().getIdLong()).queue();
-				UsrMsgUtil.sendEMessage("Invites are not allowed! "+ e.getAuthor().getAsMention(), e.getChannel());
-		    }
+			if(!e.getMember().hasPermission(Bumblebot.modPerms)) {
+				if (e.getMessage().getContentRaw().replaceAll(this.inviteRegex, "").length() < e.getMessage().getContentRaw().length()) {
+					e.getChannel().deleteMessageById(e.getMessage().getIdLong()).queue();
+					UsrMsgUtil.sendEMessage("Invites are not allowed! " + e.getAuthor().getAsMention(), e.getChannel());
+				}
+			}
 		}
 	}
 	
 	public void onGuildMessageUpdate(GuildMessageUpdateEvent e) {
 		if(e.getGuild().getId().equals(ConfigUtil.getServerId())) {
-			if(e.getMessage().getContentRaw().replaceAll(this.inviteRegex, "").length() < e.getMessage().getContentRaw().length()) {
-				e.getChannel().deleteMessageById(e.getMessage().getIdLong()).queue();
-				UsrMsgUtil.sendEMessage("Invites are not allowed! " + e.getAuthor().getAsMention() + " ... and nice try :D", e.getChannel());
+			if(!e.getMember().hasPermission(Bumblebot.modPerms)) {
+				if(e.getMessage().getContentRaw().replaceAll(this.inviteRegex, "").length() < e.getMessage().getContentRaw().length()) {
+					UsrMsgUtil.sendEMessage("Invites are not allowed! " + e.getAuthor().getAsMention() + " ... and nice try :D", e.getChannel());
+					e.getChannel().deleteMessageById(e.getMessage().getIdLong()).queue();
+				}
 		    }
 		}
 	}
